+++ packages/ariakit-solid-core/src/utils/misc.ts
@@ -1,5 +1,4 @@
 import type { AnyObject } from "@ariakit/core/utils/types";
-import { combineProps } from "@solid-primitives/props";
 import { type MaybeAccessor, access } from "@solid-primitives/utils";
 import {
   type Accessor,
@@ -11,7 +10,6 @@ import {
   splitProps,
   untrack,
 } from "solid-js";
-import type { WrapInstance, WrapInstanceValue } from "./types.ts";
 
 /**
  * Creates a stable accessor. Useful when creating derived accessors that
@@ -69,49 +67,6 @@ export function extractPropsWithDefaults<
     mergeProps(defaults, own),
     rest,
   ] as unknown as ExtractPropsWithDefaultsReturn<P, D>;
-}
-
-/**
- * Generates a unique ID.
- */
-export function createId(
-  defaultId?: MaybeAccessor<string | undefined>,
-): Accessor<string> {
-  const id = createUniqueId();
-  return () => access(defaultId) ?? id;
-}
-
-/**
- * Returns the tag name by parsing an element.
- * @example
- * function Component(props) {
- *   const [ref, setRef] = createSignal();
- *   const tagName = extractTagName(ref, "button"); // () => "div"
- *   return <div ref={setRef} {...props} />;
- * }
- */
-export function extractTagName(
-  refOrElement?: MaybeAccessor<HTMLElement | undefined>,
-  fallback?: ValidComponent,
-) {
-  return () => {
-    const element = access(refOrElement);
-    return (
-      element?.tagName.toLowerCase() ??
-      (typeof fallback === "string" ? fallback : undefined)
-    );
-  };
-}
-
-/**
- * Returns props with an additional `wrapInstance` prop.
- */
-export function wrapInstance<P, Q = P & { wrapInstance: WrapInstance }>(
-  props: P & { wrapInstance?: WrapInstance },
-  element: WrapInstanceValue,
-): Q {
-  const wrapInstance = [...(props.wrapInstance ?? []), element];
-  return combineProps(props, { wrapInstance }) as Q;
 }
 
 /**
@@ -211,5 +166,37 @@ export function createRef<T>(initialValue?: any): any {
     },
     set,
     reset: () => set(initialValue),
+  };
+}
+
+/**
+ * Generates a unique ID.
+ */
+export function createId(
+  defaultId?: MaybeAccessor<string | undefined>,
+): Accessor<string> {
+  const id = createUniqueId();
+  return () => access(defaultId) ?? id;
+}
+
+/**
+ * Returns the tag name by parsing an element.
+ * @example
+ * function Component(props) {
+ *   const [ref, setRef] = createSignal();
+ *   const tagName = extractTagName(ref, "button"); // () => "div"
+ *   return <div ref={setRef} {...props} />;
+ * }
+ */
+export function extractTagName(
+  refOrElement?: MaybeAccessor<HTMLElement | undefined>,
+  fallback?: ValidComponent,
+) {
+  return () => {
+    const element = access(refOrElement);
+    return (
+      element?.tagName.toLowerCase() ??
+      (typeof fallback === "string" ? fallback : undefined)
+    );
   };
 }
