+++ packages/ariakit-solid-core/src/utils/system.tsx
@@ -1,4 +1,5 @@
 import type { AnyObject, EmptyObject } from "@ariakit/core/utils/types";
+import { combineProps } from "@solid-primitives/props";
 import { type ValidComponent, splitProps } from "solid-js";
 import { Dynamic } from "solid-js/web";
 import {
@@ -6,7 +7,14 @@ import {
   type ExtractPropsWithDefaultsRestProps,
   extractPropsWithDefaults,
 } from "./misc.ts";
-import type { HTMLProps, Hook, Options, Props } from "./types.ts";
+import type {
+  HTMLProps,
+  Hook,
+  Options,
+  Props,
+  WrapInstance,
+  WrapInstanceValue,
+} from "./types.ts";
 
 /**
  * Creates a Solid component instance that supports the `render` and
@@ -38,6 +46,17 @@ export function createInstance(
     }
   }
   return tree();
+}
+
+/**
+ * Returns props with an additional `wrapInstance` prop.
+ */
+export function wrapInstance<P, Q = P & { wrapInstance: WrapInstance }>(
+  props: P & { wrapInstance?: WrapInstance },
+  element: WrapInstanceValue,
+): Q {
+  const wrapInstance = [...(props.wrapInstance ?? []), element];
+  return combineProps(props, { wrapInstance }) as Q;
 }
 
 /**
